Refactor: Merge 이후 Member 관련 네이밍/타입 개선
- provider, role enum 변경
- 변수 네이밍 정리
