+++ src/main/java/com/itcen/censquare/domain/auth/oauth/CustomUserDetails.java
@@ -20,7 +20,7 @@ public class CustomUserDetails implements UserDetails {
   @Override
   public Collection<? extends GrantedAuthority> getAuthorities() {
 
-    return Collections.singleton(new SimpleGrantedAuthority(member.getRole().getAuthority()));
+    return Collections.singleton(new SimpleGrantedAuthority(member.getRole().getText()));
   }
 
   /**
