+++ src/main/java/com/itcen/censquare/domain/member/dto/MemberResDto.java
@@ -1,20 +1,15 @@
 package com.itcen.censquare.domain.member.dto;
 
-import com.itcen.censquare.domain.member.entity.Member;
 import com.itcen.censquare.domain.member.entity.Role;
 import lombok.Getter;
+import lombok.RequiredArgsConstructor;
 
 @Getter
+@RequiredArgsConstructor
 public class MemberResDto {
 
   private final Long id;
   private final String provider;
   private final Role role;
-
-  public MemberResDto(Member member) {
-    this.id = member.getId();
-    this.provider = member.getProvider();
-    this.role = member.getRole();
-  }
 
 }
