+++ src/main/java/com/itcen/censquare/domain/member/entity/enums/Provider.java
@@ -6,7 +6,7 @@ import lombok.RequiredArgsConstructor;
 @Getter
 @RequiredArgsConstructor
 public enum Provider {
-  KAKAO("카카오");
+  KAKAO("kakao");
 
   private final String text;
 }
