+++ src/main/java/com/itcen/censquare/domain/member/entity/enums/Role.java
@@ -6,7 +6,7 @@ import lombok.RequiredArgsConstructor;
 @Getter
 @RequiredArgsConstructor
 public enum Role {
-  USER("사용자"),
-  ADMIN("관리자");
+  USER("ROLE_USER"),
+  ADMIN("ROLE_ADMIN");
   private final String text;
 }
