+++ src/main/java/com/itcen/censquare/domain/member/repository/MemberRepository.java
@@ -1,6 +1,7 @@
 package com.itcen.censquare.domain.member.repository;
 
 import com.itcen.censquare.domain.member.entity.Member;
+import com.itcen.censquare.domain.member.entity.enums.Provider;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
@@ -8,6 +9,6 @@ import org.springframework.stereotype.Repository;
 @Repository
 public interface MemberRepository extends JpaRepository<Member, Long> {
 
-  Optional<Member> findByOauthIdAndProvider(String oauthId, String provider);
+  Optional<Member> findByOauthIdAndProvider(String oauthId, Provider provider);
 
 }
